Color 이슈 및 iOS 26 대응
